backend/requirements.txt
fastapi
uvicorn
python-dotenv
psycopg2-binary
sqlalchemy
alembic
python-jose[cryptography]
authlib
httpx
